handle offline CPUs
So far the code had trouble dealing with the situation of offline
CPUs.  In such a scenario the lock_on_cpu failed and thus the whole
tracing was shut down since an error was detected there and thus
nthreads_running != ncpus.

This change introduces a new thread_status Th_offline, which is kind
of an acceptable error status that is set on the thread when it could
not be started due to the CPU being offline.

To achieve that we read /proc/stat on startup and record all online
CPUs in a CPU set, such that we later in the thread can determine
whether the relevant CPU is online or leave early with thread_status
Th_offline otherwise.  For getting the application up Th_offline is
counted towards success, such that the overall application does not
fail due to offline CPUs.

Signed-off-by: Robert Schiele <rschiele@gmail.com>